+++ crates/bevy_immediate_floating_ui/src/floating_ui_ordering_plugin.rs
@@ -2,7 +2,7 @@ use bevy_ecs::{
     component::Component, entity::Entity, hierarchy::ChildOf, observer::On,
     schedule::IntoScheduleConfigs, system::Query,
 };
-use bevy_picking::events::{Pointer, Press};
+use bevy_picking::events::PointerPress;
 use bevy_ui::{GlobalZIndex, UiSystems};
 
 /// Logic to handle UI layer Z ordering
@@ -12,7 +12,7 @@ impl bevy_app::Plugin for FloatingUiOrderingPlugin {
     fn build(&self, app: &mut bevy_app::App) {
         app.add_systems(
             bevy_app::PostUpdate,
-            update_ui_layer_order.before(UiSystems::Prepare),
+            update_ui_layer_order.before_weak(UiSystems::Prepare),
         );
 
         app.add_observer(window_on_focus);
@@ -39,7 +39,7 @@ impl Default for UiBringForward {
 }
 
 fn window_on_focus(
-    pointer: On<Pointer<Press>>,
+    pointer: On<PointerPress>,
     mut forward: Query<&mut UiBringForward>,
     child_of: Query<&ChildOf>,
 ) {