+++ crates/bevy_immediate_ui/src/base.rs
@@ -9,7 +9,7 @@ impl ImmCapability for CapabilityUiBase {
     fn build<Cap: CapSet>(app: &mut bevy_app::App, cap_req: &mut ImmCapAccessRequests<Cap>) {
         app.configure_sets(
             bevy_app::PostUpdate,
-            ImmediateSystemSet::<Cap>::default().before(bevy_ui::UiSystems::Prepare),
+            ImmediateSystemSet::<Cap>::default().before_weak(bevy_ui::UiSystems::Prepare),
         );
 
         let _ = cap_req;